implemented process command line reading in Windows x64; x64 target added to all projects; removed various 64-bit build warnings

+++ src/db/dbdrv/odbc/odbc.cpp
@@ -625,7 +625,7 @@ extern "C" LONG EXPORT DrvGetFieldLengthAsync(ODBCDRV_ASYNC_QUERY_RESULT *pResul
 		                        temp, 0, &dataSize);
 			if ((rc == SQL_SUCCESS) || (rc == SQL_SUCCESS_WITH_INFO))
 			{
-				nLen = dataSize;
+				nLen = (LONG)dataSize;
 			}
 		}
    }
